I updated dot-ssh to be more declarative.
I think this is a good pattern.
There are two parameters:  configuration and resources.
most of the configuration is in configuration.
resources is lightly used.
it is only used for RESOURCE.

For most everything it is done exclusively in configuration.
However for a resource.
First we define the resource in the resources.
Then we define the configuration as a lambda that returns a string.

The resource is defines a directory.
The configuration defines a file in the directory.